+++ src/module/data/pseudo-documents/message-parts/saving-throw.mjs
@@ -20,6 +20,7 @@ export default class SavingThrowPart extends BaseMessagePart {
 
   /** @inheritdoc */
   static ACTIONS = {
+    ...super.ACTIONS,
     heroToken: this.#heroToken,
   };